RAG BSAN6200 Final project/requirements.txt.txt
streamlit>=1.24.0
pdfplumber==0.11.6
numpy==2.2.5
faiss-cpu==1.11.0
rank-bm25==0.2.2
sentence-transformers==4.1.0
transformers==4.51.3
torch==2.7.0
